refactor float in bigdecimal

+++ src/hashcode2020/Library.java
@@ -1,5 +1,6 @@
 package hashcode2020;
 
+import java.math.BigDecimal;
 import java.util.*;
 
 import static java.util.Map.Entry.comparingByValue;
@@ -60,8 +61,8 @@ public class Library {
                 '}';
     }
 
-    public float getLibraryScore(int numDays, int signupStartDay, Set<Integer> books){
-        float libraryScore=strategieDiScoringLibrerie.getLibraryScore(numDays, signupStartDay, books);
+    public BigDecimal getLibraryScore(int numDays, int signupStartDay, Set<Integer> books){
+        BigDecimal libraryScore=strategieDiScoringLibrerie.getLibraryScore(numDays, signupStartDay, books);
         if(Utility.getDebug()){
             System.out.println("sto valutando la libreria "+idLibrary+" il suo score è "+ libraryScore);
         }
